Move settings classes to api/config and document config placement rule
SecretStorageSettings, AzureSettings, and ChronicleMapSettings now live in
api/config/ alongside KinoticProperties. Added config placement rule to
root CLAUDE.md: api/config/ for user-facing properties, internal/config/
for Spring @Configuration wiring.
